Factor out RequestScenario and Response from test/java/integration
