+++ test/java/integration/AndroidEnvoyExplicitFlowTest.java
@@ -18,6 +18,8 @@ import io.envoyproxy.envoymobile.Stream;
 import io.envoyproxy.envoymobile.StreamIntel;
 import io.envoyproxy.envoymobile.UpstreamHttpProtocol;
 import io.envoyproxy.envoymobile.engine.AndroidJniLibrary;
+import io.envoyproxy.envoymobile.engine.testing.RequestScenario;
+import io.envoyproxy.envoymobile.engine.testing.Response;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.ByteBuffer;
@@ -345,6 +347,7 @@ public class AndroidEnvoyExplicitFlowTest {
     });
     mockWebServer.start();
     RequestScenario requestScenario = new RequestScenario()
+                                          .closeBodyStream()
                                           .setHttpMethod(RequestMethod.POST)
                                           .setUrl(mockWebServer.url("post/flowers").toString())
                                           .addBody("This is the first part of by body. ")
@@ -369,6 +372,7 @@ public class AndroidEnvoyExplicitFlowTest {
     });
     mockWebServer.start();
     RequestScenario requestScenario = new RequestScenario()
+                                          .closeBodyStream()
                                           .setHttpMethod(RequestMethod.PUT)
                                           .setUrl(mockWebServer.url("put/flowers").toString());
     byte[] buf = new byte[10_000];
@@ -544,171 +548,7 @@ public class AndroidEnvoyExplicitFlowTest {
     if (chunkIterator.hasNext()) {
       stream.sendData(chunkIterator.next());
     } else {
-      stream.close(requestScenario.getClosingBodyChunk());
-    }
-  }
-
-  private static class RequestScenario {
-    private URL url;
-    private RequestMethod method = null;
-    private final List<ByteBuffer> bodyChunks = new ArrayList<>();
-    private final List<Map.Entry<String, String>> headers = new ArrayList<>();
-    private int responseBufferSize = 1000;
-    private boolean cancelOnResponseHeaders = false;
-    private int cancelUploadOnChunk = -1;
-    private boolean useDirectExecutor = false;
-    private boolean waitOnReadData = false;
-
-    RequestHeaders getHeaders() {
-      RequestHeadersBuilder requestHeadersBuilder =
-          new RequestHeadersBuilder(method, url.getProtocol(), url.getAuthority(), url.getPath());
-      headers.forEach(entry -> requestHeadersBuilder.add(entry.getKey(), entry.getValue()));
-      // HTTP1 is the only way to send HTTP requests (not HTTPS)
-      return requestHeadersBuilder.addUpstreamHttpProtocol(UpstreamHttpProtocol.HTTP1).build();
-    }
-
-    List<ByteBuffer> getBodyChunks() {
-      return Collections.unmodifiableList(
-          bodyChunks.subList(0, Math.max(bodyChunks.size() - 1, 0)));
-    }
-
-    ByteBuffer getClosingBodyChunk() { return bodyChunks.get(bodyChunks.size() - 1); }
-
-    boolean hasBody() { return !bodyChunks.isEmpty(); }
-
-    RequestScenario setHttpMethod(RequestMethod requestMethod) {
-      this.method = requestMethod;
-      return this;
-    }
-
-    RequestScenario setUrl(String url) throws MalformedURLException {
-      this.url = new URL(url);
-      return this;
-    }
-
-    RequestScenario addBody(byte[] requestBodyChunk) {
-      ByteBuffer byteBuffer = ByteBuffer.wrap(requestBodyChunk);
-      bodyChunks.add(byteBuffer);
-      return this;
-    }
-
-    RequestScenario addBody(String requestBodyChunk) {
-      return addBody(requestBodyChunk.getBytes());
-    }
-
-    RequestScenario addHeader(String key, String value) {
-      headers.add(new SimpleImmutableEntry<>(key, value));
-      return this;
-    }
-
-    RequestScenario setResponseBufferSize(int responseBufferSize) {
-      this.responseBufferSize = responseBufferSize;
-      return this;
-    }
-
-    RequestScenario cancelOnResponseHeaders() {
-      this.cancelOnResponseHeaders = true;
-      return this;
-    }
-
-    public RequestScenario cancelUploadOnChunk(int chunkNo) {
-      this.cancelUploadOnChunk = chunkNo;
-      return this;
-    }
-
-    public RequestScenario useDirectExecutor() {
-      this.useDirectExecutor = true;
-      return this;
-    }
-
-    public RequestScenario waitOnReadData() {
-      this.waitOnReadData = true;
-      return this;
-    }
-  }
-
-  private static class Response {
-
-    private final AtomicReference<ResponseHeaders> headers = new AtomicReference<>();
-    private final AtomicReference<ResponseTrailers> trailers = new AtomicReference<>();
-    private final AtomicReference<EnvoyError> envoyError = new AtomicReference<>();
-    private final List<StreamIntel> streamIntels = new ArrayList<>();
-    private final AtomicReference<FinalStreamIntel> finalStreamIntel = new AtomicReference<>();
-    private final List<ByteBuffer> bodies = new ArrayList<>();
-    private final AtomicBoolean cancelled = new AtomicBoolean(false);
-    private final AtomicReference<AssertionError> assertionError = new AtomicReference<>();
-    private int requestChunkSent = 0;
-
-    void setHeaders(ResponseHeaders headers) {
-      if (!this.headers.compareAndSet(null, headers)) {
-        assertionError.compareAndSet(
-            null, new AssertionError("setOnResponseHeaders called more than once."));
-      }
-    }
-
-    void addBody(ByteBuffer body) { bodies.add(body); }
-
-    void setTrailers(ResponseTrailers trailers) {
-      if (!this.trailers.compareAndSet(null, trailers)) {
-        assertionError.compareAndSet(
-            null, new AssertionError("setOnResponseTrailers called more than once."));
-      }
-    }
-
-    void addStreamIntel(StreamIntel streamIntel) { streamIntels.add(streamIntel); }
-
-    void setEnvoyError(EnvoyError envoyError) {
-      if (!this.envoyError.compareAndSet(null, envoyError)) {
-        assertionError.compareAndSet(null, new AssertionError("setOnError called more than once."));
-      }
-    }
-
-    void setFinalStreamIntel(FinalStreamIntel finalStreamIntel) {
-      if (!this.finalStreamIntel.compareAndSet(null, finalStreamIntel)) {
-        assertionError.compareAndSet(
-            null, new AssertionError("setFinalStreamIntel called more than once."));
-      }
-    }
-
-    void setCancelled() {
-      if (!cancelled.compareAndSet(false, true)) {
-        assertionError.compareAndSet(null,
-                                     new AssertionError("setOnCancel called more than once."));
-      }
-    }
-
-    List<StreamIntel> getStreamIntels() { return streamIntels; }
-
-    EnvoyError getEnvoyError() { return envoyError.get(); }
-
-    FinalStreamIntel getFinalStreamIntel() { return finalStreamIntel.get(); }
-
-    ResponseHeaders getHeaders() { return headers.get(); }
-
-    ResponseTrailers getTrailers() { return trailers.get(); }
-
-    boolean isCancelled() { return cancelled.get(); }
-
-    int getRequestChunkSent() { return requestChunkSent; }
-
-    String getBodyAsString() {
-      int totalSize = bodies.stream().mapToInt(ByteBuffer::limit).sum();
-      byte[] body = new byte[totalSize];
-      int pos = 0;
-      for (ByteBuffer buffer : bodies) {
-        int bytesToRead = buffer.limit();
-        buffer.get(body, pos, bytesToRead);
-        pos += bytesToRead;
-      }
-      return new String(body);
-    }
-
-    int getNbResponseChunks() { return bodies.size(); }
-
-    void throwAssertionErrorIfAny() {
-      if (assertionError.get() != null) {
-        throw assertionError.get();
-      }
+      requestScenario.getClosingBodyChunk().ifPresent(stream::close);
     }
   }
 }
