+++ test/java/integration/BUILD
@@ -38,6 +38,7 @@ envoy_mobile_android_test(
     ],
     deps = [
         "//library/kotlin/io/envoyproxy/envoymobile:envoy_lib",
+        "//test/java/io/envoyproxy/envoymobile/engine/testing:testing"
     ],
 )
 
@@ -56,5 +57,6 @@ envoy_mobile_android_test(
     ],
     deps = [
         "//library/kotlin/io/envoyproxy/envoymobile:envoy_lib",
+        "//test/java/io/envoyproxy/envoymobile/engine/testing:testing"
     ],
 )
